refactor(app): move file validation
